update install script

+++ install-aliases.sh
@@ -6,9 +6,19 @@ INSTALL_PATH=~/.gitaliases
 mkdir -p $INSTALL_PATH
 
 for alias in $ALL_ALIASES; do 
-  command="!sh ${INSTALL_PATH}/${alias}.sh"
-  echo "install global alias '$alias' - (command: '$command')"
-  cp -f aliases/$alias.sh $INSTALL_PATH/
-  git config --global --unset-all alias.$alias
-  git config --global alias.$alias "${command}"
+  exists=$(git config --global --get-regexp ^alias.$alias)
+  if [ ! -z "$exists" ]; then
+    read -p "Alias '$alias' already exists, do you want to overwrite it? (Y/n) " yn
+  fi
+  case $yn in
+    ""|[yY] ) 
+      command="!sh ${INSTALL_PATH}/${alias}.sh";
+      cp -f aliases/$alias.sh $INSTALL_PATH/;
+      git config --global --unset-all alias.$alias;
+      git config --global alias.$alias "${command}";
+      echo "installed global alias '$alias' - (command: '$command')";;
+    * ) 
+      echo "skipped '$alias'";;
+  esac
 done;
+
